+++ CHANGELOG.md
@@ -3,3 +3,4 @@ v0.5
 - added api reference link above block code editor
 - added changelog
 - added `que.whenTime`
+- errors now logged to the console's of blocks